# backend/swagger.json
{
	"info": {
		"title": "Gym Buddy REST API",
		"version": "0.0.1",
		"description": "API for gym buddy frontend"
	},
	"servers": [
		{
			"url": "http://localhost:5312"
		}
	],
	"swagger": "2.0",
	"paths": {
		"/account/test": {
			"get": {
				"produces": [
					"application/json"
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/definitions/schemas/Test"
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/schemas/Test2"
								}
							}
						}
					}
				}
			}
		},
		"/account/login": {
			"post": {
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/definitions/requests/UsernamePassword"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "json web token",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"token": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid request"
					},
					"401": {
						"description": "Unauthorised"
					}
				}
			}
		},
		"/account/create": {
			"post": {
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/definitions/requests/UsernamePassword"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "User created",
						"type": "string",
						"content": {
							"text/plain": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request",
						"content": {
							"text/plain": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		}
	},
	"definitions": {
		"schemas": {
			"Test": {
				"type": "object",
				"properties": {
					"property1": {
						"type": "string"
					},
					"property2": {
						"type": "string"
					}
				}
			},
			"Test2": {
				"type": "object",
				"properties": {
					"property3": {
						"type": "string"
					},
					"property4": {
						"type": "string"
					}
				}
			}
		},
		"requests": {
			"UsernamePassword": {
				"type": "object",
				"properties": {
					"username": {
						"type": "string",
						"required": true
					},
					"password": {
						"type": "string",
						"required": true
					}
				}
			}
		}
	},
	"responses": {},
	"parameters": {},
	"securityDefinitions": {},
	"tags": []
}